floorl() -> floor() so code will compile on IBM p690.
git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/trunk@240 46abec54-52ec-0310-8d7b-a975267b85f2

+++ cmc_io.c
@@ -593,8 +593,8 @@ int parser(int argc, char *argv[], gsl_rng *r)
 		N_BIN_DIM = 2 + clus.N_BINARY;
 
 		/* safety factors, so we don't have to worry about memory management/garbage collection */
-		N_STAR_DIM = floorl(1.5 * ((double) N_STAR_DIM));
-		N_BIN_DIM = floorl(1.5 * ((double) N_BIN_DIM));
+		N_STAR_DIM = (long) floor(1.5 * ((double) N_STAR_DIM));
+		N_BIN_DIM = (long) floor(1.5 * ((double) N_BIN_DIM));
 
 		/*********************************************/
 		/* allocation of memory for global variables */
